argus1/requirements.txt
# ARGUS-1 Dependencies

# Core
requests>=2.28.0
feedparser>=6.0.0

# Claude API
anthropic>=0.18.0

# Database
supabase>=2.0.0

# API Server
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0

# Scheduler (optional - only needed for local scheduling)
apscheduler>=3.10.0

# Development
python-dotenv>=1.0.0
